Extract column definitions and improve formatting
